refactor(mainwindow): split UI and attachments by domain
